Update schemas pay, keyreg, keyregnontx
