+++ lib/algorand.encoder.spec.ts
@@ -92,143 +92,6 @@ describe("Algorand Encoding", () => {
 		expect(encodedSignedTransaction).toEqual(msgpack.encode(signedTxn, { sortKeys: true }))
 	})
 
-	it("(OK) Craft Pay Transaction", async () => {
-		// from algorand address
-		const from: string = algoEncoder.encodeAddress(Buffer.from(randomBytes(32)))
-		// to algorand address
-		const to: string = algoEncoder.encodeAddress(Buffer.from(randomBytes(32)))
-
-		// note
-		const note: string = Buffer.from(randomBytes(32)).toString("base64")
-
-		// create pay transaction
-		const txn: PayTransaction = algorandCrafter
-			.pay(1000, from, to)
-			.addFirstValidRound(1000)
-			.addLastValidRound(2000)
-			.addNote(note, "base64")
-			.addFee(1000)
-			.get()
-
-		expect(txn).toBeDefined()
-		expect(txn).toBeInstanceOf(PayTransaction)
-		expect(txn).toEqual({
-			rcv: algoEncoder.decodeAddress(to),
-			snd: algoEncoder.decodeAddress(from),
-			amt: 1000,
-			fv: 1000,
-			lv: 2000,
-			gen: genesisId,
-			gh: new Uint8Array(Buffer.from(genesisHash, "base64")),
-			note: Buffer.from(note, "base64"),
-			fee: 1000,
-			type: "pay",
-		})
-	})
-
-	it("(OK) Craft Keyreg change-online transaction", async () => {
-		// from algorand address
-		const from: string = algoEncoder.encodeAddress(Buffer.from(randomBytes(32)))
-
-		// note
-		const note: string = Buffer.from(randomBytes(32)).toString("base64")
-
-		// vote key
-		const voteKey: string = Buffer.from(randomBytes(32)).toString("base64")
-
-		// selection key
-		const selectionKey: string = Buffer.from(randomBytes(32)).toString("base64")
-
-		// state proof key
-		const stateProofKey: string = Buffer.from(randomBytes(64)).toString("base64")
-
-		// create keyreg transaction
-		const txn: KeyregTransaction = algorandCrafter
-			.changeOnline(from, voteKey, selectionKey, stateProofKey, 1000, 2000, 32)
-			.addFirstValidRound(1000)
-			.addLastValidRound(2000)
-			.addNote(note, "base64")
-			.addFee(1000)
-			.get()
-
-		expect(txn).toBeDefined()
-		expect(txn).toBeInstanceOf(KeyregTransaction)
-		expect(txn).toEqual({
-			snd: algoEncoder.decodeAddress(from),
-			votekey: Buffer.from(voteKey, "base64"),
-			selkey: Buffer.from(selectionKey, "base64"),
-			sprfkey: Buffer.from(stateProofKey, "base64"),
-			votefst: 1000,
-			votelst: 2000,
-			votekd: 32,
-			fv: 1000,
-			lv: 2000,
-			gh: new Uint8Array(Buffer.from(genesisHash, "base64")),
-			note: Buffer.from(note, "base64"),
-			fee: 1000,
-			type: "keyreg",
-		})
-	})
-
-	it("(OK) Craft Keyreg change-ofline transaction", async () => {
-		// from algorand address
-		const from: string = algoEncoder.encodeAddress(Buffer.from(randomBytes(32)))
-
-		// note
-		const note: string = Buffer.from(randomBytes(32)).toString("base64")
-
-		// create keyreg transaction
-		const txn: KeyregTransaction = algorandCrafter
-			.changeOffline(from)
-			.addFirstValidRound(1000)
-			.addLastValidRound(2000)
-			.addNote(note, "base64")
-			.addFee(1000)
-			.get()
-
-		expect(txn).toBeDefined()
-		expect(txn).toBeInstanceOf(KeyregTransaction)
-		expect(txn).toEqual({
-			snd: algoEncoder.decodeAddress(from),
-			fv: 1000,
-			lv: 2000,
-			gh: new Uint8Array(Buffer.from(genesisHash, "base64")),
-			note: Buffer.from(note, "base64"),
-			fee: 1000,
-			type: "keyreg",
-		})
-	})
-
-	it("(OK) Craft Keyreg non-participation transaction", async () => {
-		// from algorand address
-		const from: string = algoEncoder.encodeAddress(Buffer.from(randomBytes(32)))
-
-		// note
-		const note: string = Buffer.from(randomBytes(32)).toString("base64")
-
-		// create keyreg transaction
-		const txn: KeyregTransaction = algorandCrafter
-			.markNonParticipation(from)
-			.addFirstValidRound(1000)
-			.addLastValidRound(2000)
-			.addNote(note, "base64")
-			.addFee(1000)
-			.get()
-
-		expect(txn).toBeDefined()
-		expect(txn).toBeInstanceOf(KeyregTransaction)
-		expect(txn).toEqual({
-			snd: algoEncoder.decodeAddress(from),
-			fv: 1000,
-			lv: 2000,
-			gh: new Uint8Array(Buffer.from(genesisHash, "base64")),
-			note: Buffer.from(note, "base64"),
-			fee: 1000,
-			type: "keyreg",
-			nonpart: true,
-		})
-	})
-
 	it("(OK) encoding of pay transaction", async () => {
 		// from algorand address
 		const from: string = algoEncoder.encodeAddress(Buffer.from(randomBytes(32)))
