+++ package.json
@@ -24,6 +24,7 @@
     "typescript": "^5.2.2"
   },
   "dependencies": {
+    "ajv": "^8.12.0",
     "algo-msgpack-with-bigint": "^2.1.1",
     "hi-base32": "^0.5.1",
     "js-sha512": "^0.8.0"
@@ -37,7 +38,12 @@
     "rootDir": "lib",
     "testRegex": ".spec.ts$",
     "transform": {
-      ".+\\.(t|j)s$": [ "ts-jest", {"useEsm": true}]
+      ".+\\.(t|j)s$": [
+        "ts-jest",
+        {
+          "useEsm": true
+        }
+      ]
     },
     "coverageDirectory": "./coverage",
     "collectCoverageFrom": [
@@ -51,6 +57,8 @@
     "moduleNameMapper": {
       "(.+)\\.js": "$1"
     },
-    "extensionsToTreatAsEsm": [".ts"]
+    "extensionsToTreatAsEsm": [
+      ".ts"
+    ]
   }
 }
